add logic to start a go routine to watch for database fields

+++ parser/config.go
@@ -19,7 +19,7 @@ type Database struct {
 	Type       string   `json:"type,omitempty"`
 	User       string   `json:"user,omitempty"`
 	Password   string   `json:"password,omitempty"`
-	Threshold  float32  `json:"threshold,omitempty"`
+	Threshold  float64  `json:"threshold,omitempty"`
 	NotifyList []string `json:"notify_list"`
 	Connection *sql.DB
 }
